+++ projecte2/src/main/java/com/ra34/projecte2/repository/ProductRepository.java
@@ -15,6 +15,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
 
     //Buscar
     List<Product> findByNameContainingIgnoreCaseAndStatusTrue(String prefix);
+
     //Ordenar
     List<Product> findByStatusTrueOrderByPriceAsc();
     List<Product> findByStatusTrueOrderByPriceDesc();
@@ -22,15 +23,14 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
     //JPQL
 
     //Llista de productes que estan entre preu min i preu max ascendent
-     @Query("SELECT p FROM Product p WHERE p.price BETWEEN :min AND :max AND p.status = true ORDER BY p.price ASC")
-    Stream<Product> findByPriceRangeAsc(@Param("min") Double min, @Param("max") Double max);
+    @Query("SELECT p FROM Product p WHERE p.price BETWEEN :min AND :max AND p.status = true ORDER BY p.price ASC")
+    Stream<Product> findByPriceRangeAsc(@Param("min") Float min, @Param("max") Float max);
 
     //Llista de productes que estan entre preu min i preu max descendent
     @Query("SELECT p FROM Product p WHERE p.price BETWEEN :min AND :max AND p.status = true ORDER BY p.price DESC")
-    Stream<Product> findByPriceRangeDesc(@Param("min") Double min, @Param("max") Double max);
-    
+    Stream<Product> findByPriceRangeDesc(@Param("min") Float min, @Param("max") Float max);
+
     //Llista top 5 amb els millors preus
     @Query("SELECT p FROM Product p WHERE p.status = true ORDER BY p.rating / p.price DESC")
     Stream<Product> topByQualityPrice();
-    
 }
